+++ Runtime/ContainerManager/Attribute/AddToPermanentContainerAttribute.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Cobilas.Unity.Management.Container {
+    [AttributeUsage(AttributeTargets.Class, Inherited = false, AllowMultiple = false)]
+    public sealed class AddToPermanentContainerAttribute : AddToContainerBaseAttribute {
+        public AddToPermanentContainerAttribute(bool doNotDuplicate) : base(doNotDuplicate) { }
+
+        public AddToPermanentContainerAttribute() : base(true) { }
+    }
+}